+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.44
+### is.codion.swing.common.ui
+- AbstractControlPanelBuilder bug fixed, button builders no longer shared, stack overflow during value linking, same button linked to multiple values.
 
 ## 0.18.43
 ### is.codion.common.core